requirements.txt
# --- core runtime dependencies ---------------------------------------------
websockets>=12,<13     # async WebSocket client for exchange streams
PyYAML>=6.0            # load config/exchanges.yaml

# --- back‑compat for Python < 3.10 (union‑type syntax) ----------------------
typing-extensions>=4.0 ; python_version < "3.10"
redis>=5,<6               # redis‑py 5.x (async & pipeline support)
psutil>=5.9,<6            # memory‑usage sampling for eviction logic
